moved make pre to own file
